requirements/local.txt
-r tests.txt

django-debug-toolbar==3.8.1
ipython==8.9.0
